+++ src/azure-cli/azure/cli/command_modules/netappfiles/tests/latest/test_snapshot_commands.py
@@ -7,7 +7,7 @@ from azure.cli.testsdk import ScenarioTest, ResourceGroupPreparer
 
 POOL_DEFAULT = "--service-level 'Premium' --size 4"
 VOLUME_DEFAULT = "--service-level 'Premium' --usage-threshold 100"
-LOCATION = "westcentralus"
+LOCATION = "westus2"
 
 # No tidy up of tests required. The resource group is automatically removed
 
@@ -23,7 +23,7 @@ class AzureNetAppFilesSnapshotServiceScenarioTest(ScenarioTest):
 
     def create_volume(self, account_name, pool_name, volume_name1, rg, tags=None):
         vnet_name = self.create_random_name(prefix='cli-vnet-', length=24)
-        creation_token = volume_name1
+        file_path = volume_name1  # creation_token
         vnet_name = "cli-vnet-lefr-02"
         subnet_name = "cli-subnet-lefr-02"
         tag = "--tags %s" % tags if tags is not None else ""
@@ -31,11 +31,11 @@ class AzureNetAppFilesSnapshotServiceScenarioTest(ScenarioTest):
         self.setup_vnet(rg, vnet_name, subnet_name)
         self.cmd("netappfiles account create -g %s -a '%s' -l %s" % (rg, account_name, LOCATION)).get_output_in_json()
         self.cmd("netappfiles pool create -g %s -a %s -p %s -l %s %s %s" % (rg, account_name, pool_name, LOCATION, POOL_DEFAULT, tag)).get_output_in_json()
-        volume1 = self.cmd("netappfiles volume create --resource-group %s --account-name %s --pool-name %s --volume-name %s -l %s %s --creation-token %s --vnet %s --subnet %s %s" % (rg, account_name, pool_name, volume_name1, LOCATION, VOLUME_DEFAULT, creation_token, vnet_name, subnet_name, tag)).get_output_in_json()
+        volume1 = self.cmd("netappfiles volume create --resource-group %s --account-name %s --pool-name %s --volume-name %s -l %s %s --file-path %s --vnet %s --subnet %s %s" % (rg, account_name, pool_name, volume_name1, LOCATION, VOLUME_DEFAULT, file_path, vnet_name, subnet_name, tag)).get_output_in_json()
 
         return volume1
 
-    @ResourceGroupPreparer()
+    @ResourceGroupPreparer(name_prefix='cli_netappfiles_test_snapshot_')
     def test_create_delete_snapshots(self):
         account_name = self.create_random_name(prefix='cli-acc-', length=24)
         pool_name = self.create_random_name(prefix='cli-pool-', length=24)
@@ -46,6 +46,8 @@ class AzureNetAppFilesSnapshotServiceScenarioTest(ScenarioTest):
         volume = self.create_volume(account_name, pool_name, volume_name, rg)
         snapshot = self.cmd("az netappfiles snapshot create -g %s -a %s -p %s -v %s -s %s -l %s --file-system-id %s" % (rg, account_name, pool_name, volume_name, snapshot_name, LOCATION, volume['fileSystemId'])).get_output_in_json()
         assert snapshot['name'] == account_name + '/' + pool_name + '/' + volume_name + '/' + snapshot_name
+        # check the created fields is populated. Checking exact dates are a little harder due to session records
+        assert snapshot['created'] is not None
 
         snapshot_list = self.cmd("az netappfiles snapshot list --resource-group %s --account-name %s --pool-name %s --volume-name %s" % (rg, account_name, pool_name, volume_name)).get_output_in_json()
         assert len(snapshot_list) == 1
@@ -54,7 +56,7 @@ class AzureNetAppFilesSnapshotServiceScenarioTest(ScenarioTest):
         snapshot_list = self.cmd("az netappfiles snapshot list --resource-group %s --account-name %s --pool-name %s --volume-name %s" % (rg, account_name, pool_name, volume_name)).get_output_in_json()
         assert len(snapshot_list) == 0
 
-    @ResourceGroupPreparer()
+    @ResourceGroupPreparer(name_prefix='cli_netappfiles_test_snapshot_')
     def test_list_snapshots(self):
         account_name = self.create_random_name(prefix='cli-acc-', length=24)
         pool_name = self.create_random_name(prefix='cli-pool-', length=24)
@@ -68,7 +70,7 @@ class AzureNetAppFilesSnapshotServiceScenarioTest(ScenarioTest):
         snapshot_list = self.cmd("az netappfiles snapshot list -g {rg} -a %s -p %s -v %s" % (account_name, pool_name, volume_name)).get_output_in_json()
         assert len(snapshot_list) == 2
 
-    @ResourceGroupPreparer()
+    @ResourceGroupPreparer(name_prefix='cli_netappfiles_test_snapshot_')
     def test_get_snapshot(self):
         account_name = self.create_random_name(prefix='cli-acc-', length=24)
         pool_name = self.create_random_name(prefix='cli-pool-', length=24)